refactor: 테스트 객체 위치 이동
