+++ di/src/test/java/com/interface21/beans/factory/config/NoConstructorService.java
@@ -1,4 +1,4 @@
-package samples;
+package com.interface21.beans.factory.config;
 
 public class NoConstructorService {
